revursive tree 2 and 3 is only Program class
